src/database_API/requirements.txt
# src/requirements.txt

requests
mysql-connector-python
fastapi==0.95.2
uvicorn==0.22.0